pyproject.toml
[project]
name = "llm-perf"
version = "0.1.0"
description = "LLM providers performance evaluation tool"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0.1",
    "aiohttp>=3.9.3",
    "rich>=13.7.0",
    "asyncio>=3.4.3",
    "openai>=1.65.5",
]
